requirements.txt
fastapi
openai
streamlit
sqlalchemy
psycopg2-binary
mangum
python-dotenv